Fix copy folder error (#65)
* Fix copy folder error

* Fix linux musl issue

+++ whitebox/download_wbt.py
@@ -128,7 +128,9 @@ def download_wbt(linux_musl=False, reset=False, verbose=True):
             if verbose:
                 print("WhiteboxTools package directory: {}".format(pkg_dir))
 
-            zip_dir = zip_name.split(".")[0]
+            zip_dir = os.path.join(
+                pkg_dir, os.path.basename(zip_name).replace(".zip", "")
+            ).replace("musl", "amd64")
             src_dir = os.path.join(zip_dir, "WBT")
 
             if os.path.exists(src_dir):
@@ -139,12 +141,14 @@ def download_wbt(linux_musl=False, reset=False, verbose=True):
             if os.path.exists(new_plugin_dir):
                 shutil.rmtree(new_plugin_dir)
 
+            if os.path.exists(new_img_dir):
+                shutil.copytree(init_img_dir, new_img_dir)
+
+            if os.path.exists(new_plugin_dir):
+                shutil.copytree(init_plugin_dir, new_plugin_dir)
+
             if os.path.exists(zip_dir):
                 shutil.rmtree(zip_dir)
-
-            shutil.copytree(init_img_dir, new_img_dir)
-            shutil.copytree(init_plugin_dir, new_plugin_dir)
-
             exe_ext = ""  # file extension for MacOS/Linux
             if platform.system() == "Windows":
                 exe_ext = ".exe"
